Prefer CUDA parity over HW capacity for maxGridDim. (#715)
* ROCr : Set INT32_MAX for gridDim X.

* Set Max Grid Dim to be max of int32

---------

+++ projects/clr/rocclr/device/pal/paldevice.cpp
@@ -514,9 +514,9 @@ void NullDevice::fillDeviceInfo(const Pal::DeviceProperties& palProp,
   info_.maxWorkItemSizes_[2] = info_.maxWorkGroupSize_;
   info_.preferredWorkGroupSize_ = settings().preferredWorkGroupSize_;
 
-  info_.maxGridDim_[0] = std::numeric_limits<uint32_t>::max();
-  info_.maxGridDim_[1] = std::numeric_limits<uint32_t>::max();
-  info_.maxGridDim_[2] = std::numeric_limits<uint32_t>::max();
+  info_.maxGridDim_[0] = std::numeric_limits<int32_t>::max();
+  info_.maxGridDim_[1] = std::numeric_limits<uint16_t>::max();
+  info_.maxGridDim_[2] = std::numeric_limits<uint16_t>::max();
 
   info_.localMemType_ = CL_LOCAL;
   info_.localMemSize_ = palProp.gfxipProperties.shaderCore.ldsSizePerCu;
